Config, Settings: the unsupported-hardware message stays in its buffer
config_ask_unsupported() added each snprintf's return to the length,
and snprintf returns what it would have written: once the text was cut
short, the length passed the end of the buffer and the next call wrote
beyond it. Each piece now goes through config_msg_append(), which stops
at the end. stdarg.h is included whatever the logging setting.

next_free_ide_channel() counted channels in a uint8_t against an int.

+++ src/qt/qt_settings_bus_tracking.cpp
@@ -149,13 +149,13 @@ SettingsBusTracking::next_free_ide_channel()
     const int   channels = Harddrives::idePlan(owners) * 2;
 
     for (int pass = 0; pass < 2; pass++) {
-        for (uint8_t i = 0; i < channels; i++) {
+        for (int i = 0; i < channels; i++) {
             const int      element = ((i << 3) >> 6);
             const uint64_t mask    = 0xffULL << ((uint64_t) ((i << 3) & 0x3f));
             const bool     owned   = owners[i >> 1].onboard || (owners[i >> 1].device != nullptr);
 
             if (((pass == 1) || owned) && !(ide_tracking[element] & mask))
-                return i;
+                return (uint8_t) i;
         }
     }
